+++ Source/Common/Foundation/SoundSystem/Decoder/libMpg123Decoder.cpp
@@ -70,7 +70,7 @@ public:
         uint32_t totalSize = 0;
         bool finished = false;
         while (!finished && totalSize < Configuration::DesiredBufferSize) {
-            size_t toRead = std::min(StepSize, fileData.byte_size() - position);
+            size_t toRead = std::min(StepSize, static_cast<uint32_t>(fileData.byte_size() - position));
             size_t remain = std::min(StepSize, Configuration::DesiredBufferSize - totalSize);
 
             size_t done = 0;
